+++ Forge/changelog.md
@@ -1,2 +1,7 @@
 1.1.0:
 * Added Cotton, Onion, Rapeseed, Strawberry, Sugar Beet Seeds to villager house loot
+* Added Tomato Slice texture
+* Updated Unripe Tomato, Tomato, Rotten Tomato, Lettuce, Sliced Lettuce and Tomato Crop Stage 8 texture
+* Added texture for Lettuce, Onion, Rapeseed and Straberry Seeds
+* Added texture for Tomato Slice
+* Added Item model for Cotton Seed